refactor(core): remove self imports
