+++ packages/sanity/src/core/form/inputs/PortableText/hooks/usePortableTextMembers.tsx
@@ -1,12 +1,14 @@
+import {type Path} from '@sanity/types'
 import {isEqual, pathFor} from '@sanity/util/paths'
 import {createRef, type MutableRefObject, type ReactNode, useContext, useMemo, useRef} from 'react'
-import {type FormPatch, type Path, set} from 'sanity'
 import {PortableTextMemberItemsContext} from 'sanity/_singletons'
 
 import {pathToString} from '../../../../field'
 import {type FIXME} from '../../../../FIXME'
 import {FormInput} from '../../../components'
 import {isMemberArrayOfObjects} from '../../../members/object/fields/asserters'
+import {set} from '../../../patch'
+import {type FormPatch} from '../../../patch/types'
 import {type ArrayOfObjectsItemMember, type ObjectFormNode} from '../../../store'
 import {type ObjectInputProps, type PortableTextInputProps} from '../../../types'
 import {isArrayOfObjectsFieldMember, isBlockType} from '../_helpers'
