+++ packages/sanity/src/core/form/inputs/arrays/ArrayOfObjectsInput/InsertMenuMenuItems.tsx
@@ -2,9 +2,9 @@ import {InsertAboveIcon, InsertBelowIcon} from '@sanity/icons'
 import {type InsertMenuOptions} from '@sanity/insert-menu'
 import {type SchemaType} from '@sanity/types'
 import {useCallback, useMemo} from 'react'
-import {useTranslation} from 'sanity'
 
 import {MenuItem} from '../../../../../ui-components'
+import {useTranslation} from '../../../../i18n/hooks/useTranslation'
 import {useInsertMenuPopover} from './InsertMenuPopover'
 
 /**
