+++ packages/sanity/src/core/form/inputs/arrays/ArrayOfObjectsInput/InsertMenuPopover.tsx
@@ -5,9 +5,9 @@ import {
 import {type SchemaType} from '@sanity/types'
 import {useClickOutside, useGlobalKeyDown} from '@sanity/ui'
 import {useCallback, useMemo, useReducer, useState} from 'react'
-import {useTranslation} from 'sanity'
 
 import {Popover, type PopoverProps} from '../../../../../ui-components'
+import {useTranslation} from '../../../../i18n/hooks/useTranslation'
 
 type PopoverState = {open: boolean}
