+++ packages/sanity/src/core/form/studio/tree-editing/components/search/TreeEditingSearch.tsx
@@ -1,9 +1,10 @@
 import {SearchIcon} from '@sanity/icons'
+import {type Path} from '@sanity/types'
 import {Card, Flex, Popover, type PopoverProps, Stack, Text, TextInput} from '@sanity/ui'
 import {type ChangeEvent, type KeyboardEvent, useCallback, useMemo, useState} from 'react'
-import {type Path, useTranslation} from 'sanity'
 import {css, styled} from 'styled-components'
 
+import {useTranslation} from '../../../../../i18n/hooks/useTranslation'
 import {useSearchableList} from '../../hooks'
 import {type TreeEditingMenuItem} from '../../types'
 import {ITEM_HEIGHT, MAX_DISPLAYED_ITEMS} from './constants'
