+++ packages/sanity/src/core/form/studio/tree-editing/components/search/TreeEditingSearchMenu.tsx
@@ -1,15 +1,13 @@
+import {type Path} from '@sanity/types'
 import {Box, Button, Stack, Text} from '@sanity/ui'
 import {isEqual} from 'lodash'
 import {useCallback} from 'react'
-import {
-  CommandList,
-  getSchemaTypeTitle,
-  type Path,
-  SanityDefaultPreview,
-  supportsTouch,
-  useTranslation,
-} from 'sanity'
 
+import {CommandList} from '../../../../../components/commandList/CommandList'
+import {useTranslation} from '../../../../../i18n/hooks/useTranslation'
+import {SanityDefaultPreview} from '../../../../../preview/components/SanityDefaultPreview'
+import {getSchemaTypeTitle} from '../../../../../schema/helpers'
+import {supportsTouch} from '../../../../../util/supportsTouch'
 import {useValuePreviewWithFallback} from '../../hooks'
 import {type TreeEditingMenuItem} from '../../types'
 import {ITEM_HEIGHT} from './constants'
