+++ packages/sanity/src/core/form/studio/tree-editing/components/tree-menu/TreeEditingMenu.tsx
@@ -1,7 +1,7 @@
+import {type Path} from '@sanity/types'
 import {Stack} from '@sanity/ui'
 import {toString} from '@sanity/util/paths'
 import {memo} from 'react'
-import {type Path} from 'sanity'
 
 import {type TreeEditingMenuItem as TreeEditingMenuItemType} from '../../types'
 import {TreeEditingMenuItem} from './TreeEditingMenuItem'
