+++ packages/sanity/src/core/form/studio/tree-editing/components/tree-menu/TreeEditingMenuItem.tsx
@@ -1,5 +1,6 @@
 import {hues} from '@sanity/color'
 import {ChevronRightIcon, StackCompactIcon} from '@sanity/icons'
+import {type Path} from '@sanity/types'
 import {Button, Card, Flex, Stack, Text} from '@sanity/ui'
 // eslint-disable-next-line camelcase
 import {getTheme_v2} from '@sanity/ui/theme'
@@ -7,10 +8,11 @@ import {toString} from '@sanity/util/paths'
 import {isEqual} from 'lodash'
 import {useCallback, useEffect, useMemo, useState} from 'react'
 import {useTranslation} from 'react-i18next'
-import {getSchemaTypeTitle, type Path, SanityDefaultPreview} from 'sanity'
 import scrollIntoViewIfNeeded, {type StandardBehaviorOptions} from 'scroll-into-view-if-needed'
 import {css, styled} from 'styled-components'
 
+import {SanityDefaultPreview} from '../../../../../preview/components/SanityDefaultPreview'
+import {getSchemaTypeTitle} from '../../../../../schema/helpers'
 import {useValuePreviewWithFallback} from '../../hooks'
 import {type TreeEditingMenuItem as TreeEditingMenuItemType} from '../../types'
 import {isArrayItemPath} from '../../utils/build-tree-editing-state/utils'
