+++ packages/sanity/src/core/store/_legacy/document/document-pair/serverOperations/publish.test.ts
@@ -1,5 +1,5 @@
 import {beforeEach, describe, expect, it, jest} from '@jest/globals'
-import {type SanityDocument} from 'sanity'
+import {type SanityDocument} from '@sanity/types'
 
 import {createMockSanityClient} from '../../../../../../../test/mocks/mockSanityClient'
 import {type OperationArgs} from '../operations/types'
