+++ packages/sanity/src/core/studio/components/navbar/search/components/SortMenu.tsx
@@ -2,11 +2,11 @@ import {SortIcon} from '@sanity/icons'
 import {Card, Flex, Menu, MenuDivider} from '@sanity/ui'
 import {isEqual} from 'lodash'
 import {useCallback, useId, useMemo} from 'react'
-import {useWorkspace} from 'sanity'
 import {styled} from 'styled-components'
 
 import {Button, MenuButton, MenuItem} from '../../../../../../ui-components'
 import {useTranslation} from '../../../../../i18n'
+import {useWorkspace} from '../../../../workspace'
 import {useSearchState} from '../contexts/search/useSearchState'
 import {getOrderings} from '../definitions/getOrderings'
 import {type SearchOrdering} from '../types'
