+++ packages/sanity/src/core/studio/packageVersionStatus/PackageVersionStatusProvider.tsx
@@ -1,10 +1,10 @@
 import {Box, Button, Stack, useToast} from '@sanity/ui'
 import {type ReactNode, useCallback, useEffect, useRef} from 'react'
-import {SANITY_VERSION} from 'sanity'
 import semver from 'semver'
 
 import {hasSanityPackageInImportMap} from '../../environment/hasSanityPackageInImportMap'
 import {useTranslation} from '../../i18n'
+import {SANITY_VERSION} from '../../version'
 import {checkForLatestVersions} from './checkForLatestVersions'
 
 // How often to to check last timestamp. at 30 min, should fetch new version
